vt_drmfb: Fix visual artifacts on bitmaps crossing right display border
In that case vt_drmfb_bitblt_bitmap truncates bitmap width to not cross
the border and passes truncated width to cfb_imageblit that results into
calculation of wrong line offsets in source bitmap. To see that move
mouse pointer to right display edge and slightly touch it.

To fix the issue pass both original and truncated width to cfb_imageblit
and use former to calculate line offsets and latter to limit width of
picture.

Sponsored by:	Future Crew LLC